requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
ffmpeg-python==0.2.0
future==0.18.2
importlib-metadata==4.8.1
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-mock==3.6.1
toml==0.10.2
typing-extensions==3.10.0.2
zipp==3.5.0
